--- networkit/cpp/algebraic/algorithms/test/AlgebraicBFSGTest.h
@@ -1,23 +1,0 @@
-/*
- * AlgebraicBFSGTest.h
- *
- *  Created on: Jun 7, 2016
- *      Author: Michael
- */
-
-#ifndef NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICBFSGTEST_H_
-#define NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICBFSGTEST_H_
-
-#include "gtest/gtest.h"
-
-namespace NetworKit {
-
-class AlgebraicBFSGTest : public testing::Test {
-public:
-	AlgebraicBFSGTest() = default;
-	virtual ~AlgebraicBFSGTest() = default;
-};
-
-} /* namespace NetworKit */
-
-#endif /* NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICBFSGTEST_H_ */
