--- networkit/cpp/algebraic/algorithms/test/AlgebraicMatchingCoarseningGTest.h
@@ -1,23 +1,0 @@
-/*
- * AlgebraicMatchingCoarseningGTest.h
- *
- *  Created on: Jul 12, 2016
- *      Author: Michael Wegner (michael.wegner@student.kit.edu)
- */
-
-#ifndef NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICMATCHINGCOARSENINGGTEST_H_
-#define NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICMATCHINGCOARSENINGGTEST_H_
-
-#include "gtest/gtest.h"
-
-namespace NetworKit {
-
-class AlgebraicMatchingCoarseningGTest : public testing::Test {
-public:
-	AlgebraicMatchingCoarseningGTest() = default;
-	virtual ~AlgebraicMatchingCoarseningGTest() = default;
-};
-
-} /* namespace NetworKit */
-
-#endif /* NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICMATCHINGCOARSENINGGTEST_H_ */
