--- networkit/cpp/algebraic/algorithms/test/AlgebraicPageRankGTest.h
@@ -1,23 +1,0 @@
-/*
- * AlgebraicPageRankGTest.h
- *
- *  Created on: Jun 20, 2016
- *      Author: Michael
- */
-
-#ifndef NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICPAGERANKGTEST_H_
-#define NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICPAGERANKGTEST_H_
-
-#include "gtest/gtest.h"
-
-namespace NetworKit {
-
-class AlgebraicPageRankGTest : public testing::Test {
-public:
-	AlgebraicPageRankGTest() = default;
-	virtual ~AlgebraicPageRankGTest() = default;
-};
-
-} /* namespace NetworKit */
-
-#endif /* NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICPAGERANKGTEST_H_ */
