--- networkit/cpp/algebraic/algorithms/test/AlgebraicSpanningEdgeCentralityGTest.h
@@ -1,23 +1,0 @@
-/*
- * AlgebraicSpanningEdgeCentralityGTest.h
- *
- *  Created on: Jul 12, 2016
- *      Author: Michael Wegner (michael.wegner@student.kit.edu)
- */
-
-#ifndef NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICSPANNINGEDGECENTRALITYGTEST_H_
-#define NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICSPANNINGEDGECENTRALITYGTEST_H_
-
-#include "gtest/gtest.h"
-
-namespace NetworKit {
-
-class AlgebraicSpanningEdgeCentralityGTest : testing::Test {
-public:
-	AlgebraicSpanningEdgeCentralityGTest() = default;
-	virtual ~AlgebraicSpanningEdgeCentralityGTest() = default;
-};
-
-} /* namespace NetworKit */
-
-#endif /* NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICSPANNINGEDGECENTRALITYGTEST_H_ */
