--- networkit/cpp/algebraic/algorithms/test/AlgebraicTriangleCountingGTest.h
@@ -1,23 +1,0 @@
-/*
- * AlgebraicTriangleCountingGTest.h
- *
- *  Created on: Jul 12, 2016
- *      Author: Michael Wegner (michael.wegner@student.kit.edu)
- */
-
-#ifndef NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICTRIANGLECOUNTINGGTEST_H_
-#define NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICTRIANGLECOUNTINGGTEST_H_
-
-#include "gtest/gtest.h"
-
-namespace NetworKit {
-
-class AlgebraicTriangleCountingGTest : public testing::Test {
-public:
-	AlgebraicTriangleCountingGTest() = default;
-	virtual ~AlgebraicTriangleCountingGTest() = default;
-};
-
-} /* namespace NetworKit */
-
-#endif /* NETWORKIT_CPP_ALGEBRAIC_ALGORITHMS_TEST_ALGEBRAICTRIANGLECOUNTINGGTEST_H_ */
